♻️ move requestNewPayment mutation in usePaymentRequests hook
