+++ frontend/src/hooks/usePaymentRequests.test.ts
@@ -85,7 +85,7 @@ const mockGetGithubUserQuery = (githubUserId: number) => ({
 });
 
 const render = (projectId: string) =>
-  renderHook(() => usePaymentRequests(projectId), {
+  renderHook(() => usePaymentRequests({ projectId }), {
     wrapper: MockedProvider,
     initialProps: {
       mocks: [
